LU-3014 utils: Fix an offset overflow in file_create()
On an x86_64 machine, creating a 3 GB ZFS-based target using a file
VDev failed like this:

  mkfs.lustre FATAL: mkfs.lustre: Unable to truncate backing store:
  Invalid argument

The error, returned by the ftruncate() call in file_create(), was due
to the "int"-type calculation for the "off_t" argument.  The byte
number of 3 GB overflowed the "int" type and became a negative
"off_t".  This patch changes file_create() to take an "__u64" size
instead of an "int" one and adds "_FILE_OFFSET_BITS=64" to the
AM_CPPFLAGS of lustre/utils, so that file VDevs larger than 2 GB can
be created on both 32-bit and 64-bit x86 architectures.

Reviewed-on: http://review.whamcloud.com/5805
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/utils/mount_utils.h
@@ -131,7 +131,7 @@ int read_local_files(struct mkfs_opts *mop);
 int is_lustre_target(struct mkfs_opts *mop);
 
 /* loopback helper functions */
-int file_create(char *path, int size);
+int file_create(char *path, __u64 size);
 int loop_format(struct mkfs_opts *mop);
 int loop_setup(struct mkfs_opts *mop);
 int loop_cleanup(struct mkfs_opts *mop);
